python-requirements.txt
fastapi
uvicornjinja2
